проблема с реализацией pytest_plugins
