+++ src/tests/functional/conftest.py
@@ -1,11 +1,11 @@
 # файл со всеми общими фикстурами для тестов.
 
+import asyncio
+
 import aiohttp
 import pytest
 import pytest_asyncio
-import asyncio
-from typing import Any
-from elasticsearch import AsyncElasticsearch, Elasticsearch
+from elasticsearch import AsyncElasticsearch
 from redis.asyncio import Redis
 
 from .settings import test_settings
@@ -64,19 +64,4 @@ def es_write_data(es_client: AsyncElasticsearch):
     return inner
 
 
-@pytest_asyncio.fixture(scope='function')
-def make_get_request(fastapi_session: aiohttp.ClientSession):
-    async def inner(endpoint: str, query_data: dict):
-        url = test_settings.service_url + f'/api/v1/{endpoint}'
-        async with fastapi_session.get(url, params=query_data) as response:
-            body = await response.json() if response.headers['Content-type'] == 'application/json' else response.text()
-            headers = response.headers
-            status = response.status
-
-            response = {
-                'body': body,
-                'headers': headers,
-                'status': status
-            }
-            return response
-    return inner
+pytest_plugins = ["tests.functional.fixtures_only_for_api"]
